+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## unreleased
+
+* `WatchPath()` replaces Watch() and WatchFlags()
+* `Triggers` constants replace FSN_CREATE, etc.
+* `Event` interface replaces FileEvent
+* `Event.Path()` replaces FileEvent.Name
+* [Feature] Option to include Hidden folders/files
+* [Feature] Option to filter file names with shell patterns
+* [Feature] Option to throttle events to avoid duplicates
+* [Feature] Option to watch folders recursively
+
 ## v0.8.12 / 2013-11-13
 
 * [API] Remove FD_SET and friends from Linux adapter